requirements.txt
# requirements.txt
prefect==2.18.1
pydantic==2.7.4
griffe==0.47.0
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-dotenv==1.0.1